requirements.txt
affine==2.4.0
attrs==25.3.0
branca==0.8.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
cmocean==4.0.3
contextily==1.6.2
contourpy==1.3.3
cycler==0.12.1
folium==0.20.0
fonttools==4.60.1
geographiclib==2.1
geopandas==1.1.1
geopy==2.4.1
haversine==2.9.0
hypothesis==6.140.0
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mercantile==1.2.1
neo4j==5.28.2
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==12.0.0
pluggy==1.6.0
psycopg==3.2.10
Pygments==2.19.2
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
requests==2.32.5
shapely==2.1.2
six==1.17.0
sortedcontainers==2.4.0
tzdata==2025.2
urllib3==2.5.0
xyzservices==2025.4.0
